refactor: split MainLayout, drop React.memo on Library and Playlists
MainLayout was a ~1050-line god component owning window chrome, a
playlist sidebar with drag-and-drop state, context menus, dialogs, a
view switcher, and the settings drawer. Every transient UI state
change (dragOverPlaylistId on every dragover, isMaximized on every
window event, contextMenu open/close, etc.) re-rendered the whole
shell, which is exactly why Library and Playlists each had to be
wrapped in React.memo as a defensive shield against the cascade.
That memo was a symptom of an architectural smell, not a real fix —
"Before You memo()" applies directly here: move state down so the
parent stops re-rendering.

This commit decomposes MainLayout into focused components, each
owning the state it needs, and lifts sidebar drawer state into
uiStore so Library and Playlists subscribe directly instead of
receiving it through props:

- WindowControls.tsx — owns isMaximized, traffic-light buttons,
  window IPC effects.
- Sidebar.tsx — owns dragOverPlaylistId, contextMenu, rename/create
  dialog state, playlists subscription, drag-and-drop handlers.
  Renders WindowControls + the persistent left drawer + portaled
  Menu/Dialog elements.
- MainContent.tsx — reads currentView from uiStore, renders
  <Library /> or <Playlists />. Its only re-render trigger is a
  view switch.
- MainLayout.tsx — thin shell (~290 lines). Loading screen, IPC
  listeners for keyboard/custom navigation events, the persistent
  Main margin animation driven by sidebarOpen, the settings slide-
  over Drawer, Player, and NotificationSystem. Effect listeners
  read store actions via getState() so dep arrays stay empty.
- uiStore.ts — adds sidebarOpen + setSidebarOpen + toggleSidebar.
  Sidebar reads them; Library and Playlists read them via selector
  in place of the drawerOpen / onDrawerToggle props that no longer
  exist.

With the parent no longer pushing renders for state Library and
Playlists don't care about, the React.memo wrappers and their
justification comments are dead weight. Both components export
their plain function components again. Library and Playlists each
re-render only when (a) view switches mount them, (b) their own
store subscriptions fire, or (c) the user actually toggles the
sidebar — every one a legitimate trigger.

All 121 e2e tests still pass against a fresh build.
